whiteboard/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
channels==4.0.0
channels-redis==4.1.0
Django==4.2.4
django-six==1.0.5
djangorestframework==3.14.0
msgpack==1.0.5
pytz==2023.3
redis==5.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
